Update networkPolicy, update page index, move files to proper dirs
