+++ src/pages/Home.tsx
@@ -20,7 +20,7 @@ import maskImage from "../images/homepage_mask.jpg";
 import threePeopleImage from "../images/homepage_three_people.jpg";
 
 import HomeNeedLeadView from "./HomeNeedLeadView";
-import Footer from "./Footer";
+import Footer from "../components/footer/Footer";
 
 const NO_MODAL = "";
 const MODAL_LEAD_OPEN = "modal_lead_open";
