+++ tsconfig.json
@@ -20,6 +20,6 @@
     "jsx": "react"
   },
   "include": [
-    "src"
+    "src", "tests/dummyData.tsx"
   ]
 }
